v3.0: Mejoras de UI/UX y correcciones del sistema
## Nuevas Funcionalidades
- Validaciones mejoradas en formularios de usuarios (solo letras en nombres, alfanumérico en usuario)
- Sistema de impresión de etiquetas con códigos de barras CODE128
- Impresión masiva de etiquetas de productos (formato A4)
- Integración de lector de código de barras en módulo de ventas
- Sistema de alertas de vencimiento en dashboard (productos vencidos y próximos a vencer)

## Correcciones
- Visualización de imágenes JPEG en productos (fallback para nombres con/sin timestamp)
- Formato de días hasta vencimiento ahora en enteros (no decimales)
- Nombres completos de clientes (nombre + apellido) en dashboard y ventas
- Código de impresión ya no visible en vista de productos (escapado de etiquetas script)
- Localización de fechas en español usando Carbon (diffForHumans)

## Mejoras Técnicas
- Carbon configurado globalmente en español (AppServiceProvider)
- Validaciones duales: frontend (HTML5 + JS) y backend (Laravel)
- Mensajes de error personalizados en español

+++ OneDrive/Documentos/3° Año TSDS/2° CUATRIMESTRE/KioscoNet/KioscoNet/app/Models/Producto.php
@@ -184,7 +184,7 @@ class Producto extends Model
     }
 
     /**
-     * Obtener días hasta vencimiento
+     * Obtener días hasta vencimiento (entero)
      */
     public function diasHastaVencimiento()
     {
@@ -192,7 +192,7 @@ class Producto extends Model
             return null;
         }
 
-        return now()->diffInDays($this->fecha_vencimiento, false);
+        return (int) now()->diffInDays($this->fecha_vencimiento, false);
     }
 
     // ==========================================
